move bandit-flag DTOs
