+++ tests/Config/ConfigurationStoreTest.php
@@ -7,8 +7,8 @@ use Eppo\Bandits\BanditReferenceIndexer;
 use Eppo\Cache\DefaultCacheFactory;
 use Eppo\Config\ConfigurationStore;
 use Eppo\DTO\Bandit\Bandit;
-use Eppo\DTO\Bandit\BanditFlagVariation;
 use Eppo\DTO\Bandit\BanditModelData;
+use Eppo\DTO\BanditFlagVariation;
 use Eppo\DTO\BanditReference;
 use Eppo\DTO\Flag;
 use Eppo\DTO\VariationType;
